requirements.txt
numpy>=1.14.5
opencv-contrib-python>=3.4.0.12
opencv-python>=3.4.0.12
tensorflow-gpu>=1.12.0
pydot>=1.3.0